requirements.txt
# FastAPI & Backend
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# LangChain
langchain==0.2.16
langchain-openai==0.2.8
langchain-anthropic==0.2.4
langchain-core==0.2.38
langchain-google-genai>=1.0.0
langchain-community>=0.2.0

# AI/ML
openai==1.50.2
openai-agents>=0.2.0
anthropic==0.8.1
tiktoken==0.5.2
litellm==1.52.0
google-generativeai>=0.3.0

# Data Processing
pandas==2.1.4
numpy==1.26.2
scipy==1.11.4

# Visualization
plotly==5.18.0
matplotlib==3.8.2
seaborn>=0.12.2
folium>=0.14.0

# Geospatial
geopandas>=0.13.0
shapely>=2.0.0

# Image Processing
Pillow>=10.0.0

# Streamlit
streamlit==1.29.0

# HTTP & Requests
requests==2.31.0
httpx==0.25.2

# Geocoding
geopy>=2.3.0

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0

# Logging
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development
black==23.12.1
flake8==6.1.0
mypy==1.7.1
isort==5.13.2

# Utilities
python-multipart==0.0.6
aiofiles==23.2.1

# Monitoring & Observability
langfuse==2.54.0

# Web Scraping (for New Product Agent)
selenium==4.15.2
webdriver-manager==4.0.1

# MCP (Model Context Protocol)
mcp>=1.0.0

